% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtw_data.R
\docType{data}
\name{wtw_data}
\alias{wtw_data}
\title{A data set of water treatment works and reported determinands}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 494199 rows and 6 columns.
}
\usage{
data(wtw_data)
}
\description{
A data set of nine water treatment works with the following determinands:
Coliforms, Colony Count 21, Colony Count 37, Colour, Conductivity, E.coli,
Free Chlorine, Odour, pH, Taste, Temperature, Total Chlorine and Turbidity
}
\keyword{datasets}
